+++ tests/tz/test_timezones.py
@@ -13,5 +13,4 @@ def test_timezones_are_loadable():
     zones = pendulum.timezones
 
     for zone in zones:
-        print(zone)
         pendulum.timezone(zone)